refactor: move/rename logic used to build grid from CSV files
