+++ powersimdata/input/tests/test_grid.py
@@ -4,8 +4,8 @@ import numpy as np
 import pandas as pd
 import pytest
 
+from powersimdata.input.converter.helpers import add_column_to_data_frame
 from powersimdata.input.grid import Grid
-from powersimdata.input.helpers import add_column_to_data_frame
 from powersimdata.input.scenario_grid import format_gencost, link
 
 INCORRECT_SOURCE = "invalid_source"
